# docs/openapi.json
{"openapi":"3.1.0","info":{"title":"RunTwi API","description":"Local-language voice fitness coach — coaching engine, phrase bank, runs, plans. Cues are spoken in Twi, Ga and Ewe.","version":"1.0.0"},"paths":{"/profiles/me":{"get":{"tags":["profiles"],"summary":"Me","operationId":"me_profiles_me_get","parameters":[{"name":"authorization","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Authorization"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProfileOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["profiles"],"summary":"Update Me","operationId":"update_me_profiles_me_put","parameters":[{"name":"authorization","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Authorization"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProfileIn"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProfileOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/plans/generate":{"post":{"tags":["plans"],"summary":"Create Generated Plan","operationId":"create_generated_plan_plans_generate_post","parameters":[{"name":"goal","in":"query","required":false,"schema":{"type":"string","default":"5k","title":"Goal"}},{"name":"weekly_km","in":"query","required":false,"schema":{"type":"number","default":20.0,"title":"Weekly Km"}},{"name":"authorization","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Authorization"}}],"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PlanOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/plans":{"post":{"tags":["plans"],"summary":"Create Plan","operationId":"create_plan_plans_post","parameters":[{"name":"authorization","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Authorization"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PlanIn"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PlanOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["plans"],"summary":"List Plans","operationId":"list_plans_plans_get","parameters":[{"name":"authorization","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Authorization"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PlanOut"},"title":"Response List Plans Plans Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/runs":{"post":{"tags":["runs"],"summary":"Sync Run","description":"Idempotent upsert keyed on client_id so retried offline syncs are safe.","operationId":"sync_run_runs_post","parameters":[{"name":"authorization","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Authorization"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RunIn"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RunOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["runs"],"summary":"List Runs","operationId":"list_runs_runs_get","parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":50,"title":"Limit"}},{"name":"authorization","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Authorization"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RunOut"},"title":"Response List Runs Runs Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/runs/stats":{"get":{"tags":["runs"],"summary":"Stats","operationId":"stats_runs_stats_get","parameters":[{"name":"authorization","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Authorization"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Response Stats Runs Stats Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/runs/{run_id}":{"get":{"tags":["runs"],"summary":"Get Run","operationId":"get_run_runs__run_id__get","parameters":[{"name":"run_id","in":"path","required":true,"schema":{"type":"string","title":"Run Id"}},{"name":"authorization","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Authorization"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RunOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/coaching/evaluate":{"post":{"tags":["coaching"],"summary":"Evaluate Tick","operationId":"evaluate_tick_coaching_evaluate_post","parameters":[{"name":"authorization","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Authorization"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EvaluateIn"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EvaluateOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/phrases/manifest":{"get":{"tags":["phrases"],"summary":"Get Manifest","operationId":"get_manifest_phrases_manifest_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Get Manifest Phrases Manifest Get"}}}}}}},"/phrases/resolve":{"get":{"tags":["phrases"],"summary":"Resolve","operationId":"resolve_phrases_resolve_get","parameters":[{"name":"cue","in":"query","required":true,"schema":{"type":"string","title":"Cue"}},{"name":"dialect","in":"query","required":false,"schema":{"type":"string","pattern":"^(twi|ga|ewe)$","default":"twi","title":"Dialect"}},{"name":"meters","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Meters"}},{"name":"km","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Km"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Response Resolve Phrases Resolve Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/health":{"get":{"tags":["meta"],"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Health Health Get"}}}}}}},"/":{"get":{"tags":["meta"],"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Root  Get"}}}}}}}},"components":{"schemas":{"EvaluateIn":{"properties":{"elapsed_s":{"type":"number","title":"Elapsed S"},"distance_m":{"type":"number","title":"Distance M"},"pace_s_per_km":{"type":"number","title":"Pace S Per Km"},"cadence_spm":{"type":"number","title":"Cadence Spm"},"mode":{"type":"string","title":"Mode","default":"free"},"dialect":{"type":"string","title":"Dialect","default":"twi"},"target_pace_s_per_km":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Target Pace S Per Km"},"target_distance_m":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Target Distance M"},"interval_phase":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Interval Phase"},"interval_phase_changed":{"type":"boolean","title":"Interval Phase Changed","default":false},"memory":{"additionalProperties":true,"type":"object","title":"Memory","default":{}}},"type":"object","required":["elapsed_s","distance_m","pace_s_per_km","cadence_spm"],"title":"EvaluateIn"},"EvaluateOut":{"properties":{"cue":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Cue"},"reason":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Reason"},"params":{"additionalProperties":true,"type":"object","title":"Params","default":{}},"audio":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Audio"},"memory":{"additionalProperties":true,"type":"object","title":"Memory"}},"type":"object","required":["cue","reason","memory"],"title":"EvaluateOut"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"PlanIn":{"properties":{"name":{"type":"string","title":"Name"},"goal":{"type":"string","pattern":"^(5k|10k|half|base)$","title":"Goal","default":"5k"},"sessions":{"items":{"$ref":"#/components/schemas/PlanSession"},"type":"array","title":"Sessions","default":[]}},"type":"object","required":["name"],"title":"PlanIn"},"PlanOut":{"properties":{"name":{"type":"string","title":"Name"},"goal":{"type":"string","pattern":"^(5k|10k|half|base)$","title":"Goal","default":"5k"},"sessions":{"items":{"$ref":"#/components/schemas/PlanSession"},"type":"array","title":"Sessions","default":[]},"id":{"type":"string","title":"Id"},"active":{"type":"boolean","title":"Active"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["name","id","active","created_at"],"title":"PlanOut"},"PlanSession":{"properties":{"day":{"type":"integer","title":"Day"},"type":{"type":"string","pattern":"^(easy|interval|tempo|long|rest)$","title":"Type","default":"easy"},"distance_km":{"type":"number","title":"Distance Km","default":0},"target_pace_s_per_km":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Target Pace S Per Km"},"notes":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Notes"}},"type":"object","required":["day"],"title":"PlanSession"},"ProfileIn":{"properties":{"display_name":{"type":"string","title":"Display Name","default":"Runner"},"dialect":{"type":"string","pattern":"^(twi|ga|ewe)$","title":"Dialect","default":"twi"},"weekly_target_km":{"type":"number","title":"Weekly Target Km","default":20.0}},"type":"object","title":"ProfileIn"},"ProfileOut":{"properties":{"display_name":{"type":"string","title":"Display Name","default":"Runner"},"dialect":{"type":"string","pattern":"^(twi|ga|ewe)$","title":"Dialect","default":"twi"},"weekly_target_km":{"type":"number","title":"Weekly Target Km","default":20.0},"id":{"type":"string","title":"Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","created_at"],"title":"ProfileOut"},"RoutePoint":{"properties":{"t":{"type":"number","title":"T"},"lat":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Lat"},"lon":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Lon"},"distance_m":{"type":"number","title":"Distance M","default":0},"pace":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Pace"},"cad":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Cad"}},"type":"object","required":["t"],"title":"RoutePoint"},"RunIn":{"properties":{"client_id":{"type":"string","title":"Client Id"},"mode":{"type":"string","pattern":"^(free|interval|tempo|long)$","title":"Mode","default":"free"},"dialect":{"type":"string","pattern":"^(twi|ga|ewe)$","title":"Dialect","default":"twi"},"plan_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Plan Id"},"target_pace_s_per_km":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Target Pace S Per Km"},"started_at":{"type":"string","format":"date-time","title":"Started At"},"ended_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Ended At"},"route":{"items":{"$ref":"#/components/schemas/RoutePoint"},"type":"array","title":"Route","default":[]},"cues":{"items":{"additionalProperties":true,"type":"object"},"type":"array","title":"Cues","default":[]}},"type":"object","required":["client_id","started_at"],"title":"RunIn"},"RunOut":{"properties":{"id":{"type":"string","title":"Id"},"mode":{"type":"string","title":"Mode"},"dialect":{"type":"string","title":"Dialect"},"distance_m":{"type":"number","title":"Distance M"},"duration_s":{"type":"number","title":"Duration S"},"avg_pace_s_per_km":{"type":"number","title":"Avg Pace S Per Km"},"avg_cadence_spm":{"type":"number","title":"Avg Cadence Spm"},"splits":{"items":{"additionalProperties":true,"type":"object"},"type":"array","title":"Splits"},"started_at":{"type":"string","format":"date-time","title":"Started At"},"ended_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Ended At"}},"type":"object","required":["id","mode","dialect","distance_m","duration_s","avg_pace_s_per_km","avg_cadence_spm","splits","started_at","ended_at"],"title":"RunOut"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}